Rename ruby files using snake_case
